requirements.txt
Flask==0.10.1
tornado==4.0.1
websocket-client==0.16.0